chore: Fork Test - Substrate Update (#1)
* draft: PSVM + shallow dependencies

* draft: Working version - without local fully forks updated

* chore: sc-cli - updated

* chore: sc-cli - ForkAware tx pool

* chore: sp-trie & sp-state-machine

* chore: FMT

* chore: Test fixes

* chore: Dependencies + format

* chore: CI new dependencies

* fix: CI - CLANG

* fix: CI - CLANG v2

* fix: CLANG v3

* fix: CI - CLANG v4

* fix: CI - CLANG v5

* fix: CI - CLANG v6

* fix: CI - all full builds update

* fix: Local runtime tests

* fix: PM wormhole check

+++ primitives/trie/src/node_header.rs
@@ -54,7 +54,7 @@ use trie_constants::EMPTY_TRIE;
 impl Encode for NodeHeader {
 	fn encode_to<T: Output + ?Sized>(&self, output: &mut T) {
 		let value: u64 = match self {
-			NodeHeader::Null => EMPTY_TRIE, // Type 0
+			NodeHeader::Null => EMPTY_TRIE as u64, // Type 0
 			NodeHeader::Branch(true, nibble_count) => {
 				(1u64 << 60) | (*nibble_count as u64) // Type 1
 			},
